[android] Extracted ping subscription service logic in a abstraction fragment
